Move caching load* methods to a mixin class.
